스와이프 삭제 시 알럿 창 추가

+++ PhotoTodo/FolderListView.swift
@@ -20,93 +20,132 @@ struct FolderListView: View {
     @Environment(\.modelContext) private var modelContext
     @State private var editMode: EditMode = .inactive
     @Query private var folders: [Folder]
-    @Query private var folderOrders: [FolderOrder]
+    
+    //폴더 삭제시
+    @State private var showingAlert = false
+    @State private var selectedFolder: Folder? = nil
+    @State private var pendingCompletion: (() -> Void)?
+    
+    //폴더 생성 및 편집 관련 sheet에서 쓰임
     @State var isShowingSheet = false
     @State var folderNameInput = ""
     @State var selectedColor: Color?
+    
+    @Query private var folderOrders: [FolderOrder]
+    
     private var basicViewType: TodoGridViewType = .singleFolder
     private var doneListViewType: TodoGridViewType = .doneList
     @AppStorage("defaultFolderID") private var defaultFolderID: String?
     
+    
     // 유저가 설정한 순서대로 폴더를 정렬해 보여줌
-       var orderedFolder: [Folder] {
-           let uuidLookup = Dictionary(grouping: folders, by: { $0.id })
-           return folderOrders.first?.uuidOrder.compactMap({ uuidLookup[$0]?.first }) ?? []
-       }
+    var orderedFolder: [Folder] {
+        let uuidLookup = Dictionary(grouping: folders, by: { $0.id })
+        return folderOrders.first?.uuidOrder.compactMap({ uuidLookup[$0]?.first }) ?? []
+    }
     
     
     var body: some View {
-            List {
-                //기본 폴더(인덱스 0에 있음) → 삭제 불가능하게 만들기 위해 따로 뺌
-                NavigationLink{
-                    TodoGridView(currentFolder: folders.count > 0 ? folders.first(where: {$0.id.uuidString == defaultFolderID}) : nil, viewType: basicViewType)
-                } label : {
-                    FolderRow(folder: folders.count > 0 ? folders.first(where: {$0.id.uuidString == defaultFolderID}) : nil, viewType: basicViewType)
-                }
-
-                
-                //기본 폴더를 제외하고는 모두 삭제 가능
-                ForEach(orderedFolder.filter({$0.id.uuidString != defaultFolderID})) { folder in
-                    ZStack{ //editMode에서 chevron을 숨기고, FolderRow의 Text의 투명도를 그대로 유지하기 위해서 ZStack을 활용했습니다.
-                        NavigationLink {
-                            TodoGridView(currentFolder: folder, viewType: basicViewType)
-                        } label: {
-                            EmptyView()
-                        }
-                        .opacity(editMode == .active ? 0 : 1)
-                        .swipeActions(content: {
-                            Button("Delete", systemImage: "trash", role:  .destructive) {
-                                deleteFolder(folder)
+        List {
+            //기본 폴더(인덱스 0에 있음) → 삭제 불가능하게 만들기 위해 따로 뺌
+            NavigationLink{
+                TodoGridView(currentFolder: folders.count > 0 ? folders.first(where: {$0.id.uuidString == defaultFolderID}) : nil, viewType: basicViewType)
+            } label : {
+                FolderRow(folder: folders.count > 0 ? folders.first(where: {$0.id.uuidString == defaultFolderID}) : nil, viewType: basicViewType)
+            }
+            
+            
+            //기본 폴더를 제외하고는 모두 삭제 가능
+            ForEach(orderedFolder.filter({$0.id.uuidString != defaultFolderID})) { folder in
+                ZStack{
+                    FolderRowView(actions: [
+                        Action(color: .red, name: "delete", systemIcon: "trash.fill", action: { completion in
+                            selectedFolder = folder
+                            pendingCompletion = completion
+                            showingAlert = true
+                        })]) {
+                            NavigationLink {
+                                TodoGridView(currentFolder: folder, viewType: basicViewType)
+                            } label: {
+                                FolderRow(folder: folder, viewType: basicViewType)
                             }
-                        })
-                        FolderRow(folder: folder, viewType: basicViewType)
-                    }
-                }
-                .onMove(perform: handleMove)
-//                .onDelete { _ in
-//                    // editMode일 때 UI가 반응하도록 하기 위해 빈 클로저를 남겼습니다.
-//                    // 실제 삭제 실행시에는 swipeActions으로 넘겨받은 클로저가 호출됩니다.
-//                }
-                //TODO: 옵션을 줘서 완료된 것(되지 않은 것)만 필터링해서 보여주기
-                //리스트 뷰의 마지막에는 완료함이 위치함
-                NavigationLink {
-                    DoneListView()
-                } label : {
-                    FolderRow(folder: nil, viewType: doneListViewType)
-                }
-            }
-            .onAppear {
-                // 기본폴더 세팅 로직
-                if defaultFolderID == nil {
-                    defaultFolderID = folders.first(where: {$0.name == "기본"})?.id.uuidString
-                }
-                
-                //folderOrder 세팅 로직
-                setFolderOrders()
-            }
-
-            .scrollContentBackground(.hidden)
-            //            .background(Color.white.edgesIgnoringSafeArea(.all))
-            .navigationBarTitle("폴더")
-            .toolbar {
-                ToolbarItem {
-                    if editMode == .inactive {
-                        Button(action: toggleShowingSheet) {
-                            Label("add a folder", systemImage: "plus")
                         }
+                        .opacity( editMode == .active ? 0 : 1)
+                        .disabled(editMode == .active)
+                    FolderRow(folder: folder, viewType: basicViewType)
+                        .padding(.leading, 16)
+                        .opacity( editMode == .active ? 1 : 0)
+                        
+                }
+                .listRowInsets(EdgeInsets(top: -5, leading: 0, bottom: -5, trailing: 0))
+
+            }
+            .onMove(perform: handleMove)
+            // .onDelete { _ in
+            //     // editMode일 때 UI가 반응하도록 하기 위해 빈 클로저를 남겼습니다.
+            //     // 실제 삭제 실행시에는 swipeActions으로 넘겨받은 클로저가 호출됩니다.
+            // }
+            //TODO: 옵션을 줘서 완료된 것(되지 않은 것)만 필터링해서 보여주기
+            //리스트 뷰의 마지막에는 완료함이 위치함
+            NavigationLink {
+                DoneListView()
+            } label : {
+                FolderRow(folder: nil, viewType: doneListViewType)
+            }
+        }
+        .animation(.default, value: folders) //폴더의 변화가 자연스럽게 반영되도록 설정
+        .onAppear {
+            // 기본폴더 세팅 로직
+            if defaultFolderID == nil {
+                defaultFolderID = folders.first(where: {$0.name == "기본"})?.id.uuidString
+            }
+            
+            //folderOrder 세팅 로직
+            setFolderOrders()
+        }
+        
+        .scrollContentBackground(.hidden)
+        //            .background(Color.white.edgesIgnoringSafeArea(.all))
+        .navigationBarTitle("폴더")
+        .toolbar {
+            ToolbarItem {
+                if editMode == .inactive {
+                    Button(action: toggleShowingSheet) {
+                        Label("add a folder", systemImage: "plus")
                     }
                 }
-
-                ToolbarItem(placement: .navigationBarTrailing) {
-                    EditButton()
-                        .frame(width: 38)
-                }
             }
-            .sheet(isPresented: $isShowingSheet, content: {
-                FolderEditView(isSheetPresented: $isShowingSheet, folderNameInput: $folderNameInput, selectedColor: $selectedColor)
-                    .presentationDetents([.medium, .large])
-            })
-            .environment(\.editMode, $editMode)
+            
+            ToolbarItem(placement: .navigationBarTrailing) {
+                EditButton()
+                    .frame(width: 38)
+            }
+        }
+        .alert(isPresented: $showingAlert) {
+            Alert(
+                title: Text("'\(selectedFolder?.name ?? "")' 폴더를 삭제하시겠습니까?"),
+                message: Text("이 폴더의 모든 사진이 삭제됩니다."),
+                primaryButton: .default(
+                    Text("취소"),
+                    action: {
+                        pendingCompletion?()
+                        pendingCompletion = nil
+                    }
+                ),
+                secondaryButton: .destructive(
+                    Text("삭제"),
+                    action: {
+                        deleteFolder(selectedFolder)
+                        pendingCompletion?()
+                        pendingCompletion = nil
+                    }
+                ))
+        }
+        .sheet(isPresented: $isShowingSheet, content: {
+            FolderEditView(isSheetPresented: $isShowingSheet, folderNameInput: $folderNameInput, selectedColor: $selectedColor)
+                .presentationDetents([.medium, .large])
+        })
+        .environment(\.editMode, $editMode)
     }
 }
 
@@ -129,7 +168,8 @@ extension FolderListView {
         }
     }
     
-    private func deleteFolder(_ folder: Folder) {
+    private func deleteFolder(_ folder: Folder?) {
+        guard let folder = folder else { return }
         if let folderOrder = folderOrders.first {
             modelContext.delete(folder)
             folderOrders.first?.uuidOrder.removeAll { $0 == folder.id }
@@ -174,6 +214,18 @@ private struct FolderRow: View {
     @Environment(\.editMode) private var editMode
     @State var folder: Folder?
     @State var viewType: TodoGridViewType
+    @Query private var todos: [Todo]
+    @State private var doneCount: Int = 0 // 뷰가 업데이트될 때마다 다시 계산하는 대신, 값을 미리 계산하여 사용
+    @State private var incompleteTodosCount: Int = 0
+    
+    var folderString: String {
+        return folder != nil ? folder!.name : viewType == .singleFolder ? "" : "완료함"
+    }
+    
+    var folderCountString: String {
+        return folder != nil ? "\(incompleteTodosCount)장" : viewType == .singleFolder ? "" : "\(doneCount)장"
+    }
+    
     @AppStorage("defaultFolderID") private var defaultFolderID: String?
     @State private var showingAlert = false
     @Query var folderOrders: [FolderOrder]
@@ -184,9 +236,24 @@ private struct FolderRow: View {
         HStack{
             Image(systemName: "folder.fill")
                 .foregroundStyle(viewType == .singleFolder ? changeStringToColor(colorName: folder != nil ? folder!.color : "folder-color/green" ) : Color("gray/gray-800"))
-            Text(folder != nil ? folder!.name : viewType == .singleFolder ? "" : "완료함")
+            Text(folderString)
             Spacer()
-            menu
+            ZStack{
+                Text(folderCountString)
+                    .foregroundColor(Color("gray/gray-500"))
+                    .opacity(editMode?.wrappedValue == .active ? 0 : 1)
+                menu
+            }
+
+        }
+        .onAppear {
+            // 뷰가 업데이트될 때마다 다시 계산하는 대신, 값을 미리 계산하여 사용
+            if folder == nil && viewType != .singleFolder {
+                doneCount = todos.filter { $0.isDone }.count
+            }
+            if folder != nil && viewType == .singleFolder {
+                incompleteTodosCount = folder!.todos.filter { !$0.isDone }.count
+            }
         }
         .alert("폴더 이름 변경", isPresented: $isRenamingFolder) {
             TextField("\(newFolderName)", text: $newFolderName)
@@ -263,7 +330,7 @@ extension FolderRow {
         .disabled(!isShowingMemu) // 인터렉션을 막기
     }
 }
-                          
+
 
 
 #Preview {
